Backend/requirements.txt
python-dotenv==1.0.0
mysql-connector-python>=8.0,<8.1
fastapi
json    
numpy
os
shutil
pytesseract
flask_cors
typing
flask
config
requests
pathlib
sentence_transformers
PIL
PyPDF2
mysql
pydantic
httpx